backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
boto3==1.37.37
pydantic==2.11.3
python-dotenv==1.1.0
websockets==12.0
